ai-service/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
opencv-python-headless
numpy
Pillow
pydantic
python-multipart
kaggle
mediapipe
scikit-learn==1.4.2
